refactored all the stuff O__________________________O
