+++ Example.h
@@ -1,0 +1,14 @@
+//
+// Created by Bastian Wegge on 10.10.17.
+//
+
+#ifndef SFML_EXAMPLE_H
+#define SFML_EXAMPLE_H
+
+
+class Example {
+
+};
+
+
+#endif //SFML_EXAMPLE_H
